requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
Arpeggio==2.0.2
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bidict==0.23.1
bitmath==1.3.3.1
bleach==6.1.0
blinker==1.7.0
cached-property==1.5.2
cbor==1.0.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chest==0.2.3
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==42.0.7
cwl-eval==1.0.12
datamaestro==1.1.0
datamaestro-text==2024.3.10
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
docstring-parser==0.15
executing==2.0.1
experimaestro==1.5.6
experimaestro-ir @ git+https://github.com/experimaestro/experimaestro-ir.git@931f5993e6cc8f69a08b68dddb2bf856d28c032c
fabric==3.2.2
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.14.0
Flask==2.3.3
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
gevent==23.9.1
gevent-websocket==0.10.1
ghp-import==2.1.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
HeapDict==1.0.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
ijson==3.2.3
inscriptis==2.5.0
install==1.3.5
invoke==2.2.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
ir-datasets==0.5.6
ir_measures==0.3.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.0
jupyterlab_widgets==3.0.10
kagglehub==0.2.5
keras==3.3.3
libclang==18.1.1
lightning-utilities==0.11.2
lxml==5.2.1
lz4==4.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matchpy==0.5.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.0.2
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.21
mkdocs-material-extensions==1.3.1
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiset==2.1.1
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
nptyping==1.4.4
numpy==1.26.4
omegaconf==2.3.0
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==24.0
paginate==0.5.6
pandas==2.2.2
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
platformdirs==4.2.0
plumbum==1.8.3
pretty-errors==1.2.25
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyautocorpus==0.1.12
pycparser==2.22
Pygments==2.17.2
pyjnius==1.6.1
pymdown-extensions==10.8.1
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-engineio==4.9.0
python-json-logger==2.0.7
python-socketio==5.11.2
python-terrier==0.10.0
pytools==2023.1.1
pytorch-lightning==2.2.4
pytrec-eval-terrier==0.5.6
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==26.0.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rpyc==6.0.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
Send2Trash==1.8.3
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.16.1
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.16.0
threadpoolctl==3.4.0
tinycss2==1.2.1
tokenizers==0.19.1
toma==1.1.0
torch==2.3.0
torchmetrics==1.4.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.2
trec-car-tools==2.6
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
typish==1.9.3
tzdata==2024.1
unlzw3==0.2.2
uri-template==1.3.0
urllib3==2.2.1
warc3-wet==0.2.3
warc3-wet-clueweb09==0.2.5
watchdog==2.3.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
wget==3.2
wheel==0.43.0
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
xpmir-rust==0.21.0
yarl==1.9.4
zlib-state==0.1.6
zope.event==5.0
zope.interface==6.3
